refactor and remove unrelated components, add pages
