+++ src/app/app.module.ts
@@ -4,15 +4,15 @@ import {HttpClientModule} from "@angular/common/http";
 import {AppRoutingModule} from "./app-routing.module";
 import {AppComponent} from './app.component';
 import {DashboardComponent} from './component/dashboard/dashboard.component';
-import {TrainingContentComponent} from './component/training-content/training-content.component';
+import {TrainingContentComponent} from './component/page/training-content/training-content.component';
 import {TrainingContentService} from "./service/TrainingContentService";
 import {GenericFilterPipe} from './pipe/generic-filter.pipe';
 import {QuestionFormComponent} from './component/question-form/question-form.component';
 import {FormsModule, ReactiveFormsModule} from "@angular/forms";
 import {QuestionAnswerNotEmptyPipe} from './pipe/question-answer/question-answer-not-empty.pipe';
-import {QuestionDisplayComponent} from './component/question-display/question-display.component';
 import {AnswerNotBlankPipe} from './pipe/answer/answer-not-blank.pipe';
 import { MessageModalComponent } from './component/message-modal/message-modal.component';
+import { RandomQuestionComponent } from './component/page/random-question/random-question.component';
 
 @NgModule({
   declarations: [
@@ -22,9 +22,9 @@ import { MessageModalComponent } from './component/message-modal/message-modal.c
     GenericFilterPipe,
     QuestionFormComponent,
     QuestionAnswerNotEmptyPipe,
-    QuestionDisplayComponent,
     AnswerNotBlankPipe,
-    MessageModalComponent
+    MessageModalComponent,
+    RandomQuestionComponent,
   ],
   imports: [
     BrowserModule,
