+++ src/app/component/page/training-content/training-content.component.html
@@ -1,9 +1,10 @@
+<!--display all questions-->
 <h2 xmlns="http://www.w3.org/1999/html">{{trainingContent?.name}}</h2>
 <button class="btn btn-warning" (click)="back()">Back</button>
 
 <div class="p-4">
   <div class="mt-4" *ngFor="let question of trainingContent?.questions">
-    <app-question-display [question]="question"></app-question-display>
+    <app-question-form [question]="question"></app-question-form>
     <hr>
   </div>
 </div>
