+++ src/app/component/page/training-content/training-content.component.spec.ts
