+++ src/app/component/page/training-content/training-content.component.ts
@@ -1,6 +1,6 @@
 import {Component, OnInit} from '@angular/core';
-import {TrainingContent} from "../../entity/TrainingContent";
-import {TrainingContentService} from "../../service/TrainingContentService";
+import {TrainingContent} from "../../../entity/TrainingContent";
+import {TrainingContentService} from "../../../service/TrainingContentService";
 import {ActivatedRoute} from "@angular/router";
 import {Location} from '@angular/common';
