--- src/app/component/question-display/question-display.component.html
@@ -1,30 +1,0 @@
-<div class="row fs-5" style=" white-space: pre-wrap;">
-  <div class="col-6">
-    <h1>{{question?.id}}</h1>
-    <div class="">{{question?.question}}</div>
-  </div>
-
-  <div class="col-6 d-flex flex-column">
-    <ng-container *ngFor="let key of objectKeys(question?.answers) | questionAnswerNotEmpty : question?.answers">
-      <div class="d-flex flex-row">
-        <div class="fs-2 fw-bold ms-3 me-4">{{key}}</div>
-        <div class="row custom-control custom-radio mb-5">
-          <div class="col-5">{{question?.answers[key]}}</div>
-        </div>
-      </div>
-    </ng-container>
-  </div>
-</div>
-
-<button class="btn btn-primary" type="button"
-        [attr.data-bs-toggle]="'collapse'" [attr.data-bs-target]="'#ans' + question?.id"
-        [attr.aria-expanded]="false">See answer
-</button>
-
-<div class="collapse" id="ans{{question?.id}}">
-  <div>Answer: <h2>{{question?.answer}}</h2></div>
-  <div class="card card-body">
-    {{question?.explanation}}
-  </div>
-</div>
-
