+++ src/app/component/question-form/question-form.component.html
@@ -1,6 +1,4 @@
 <div class="p-4">
-  <button class="btn btn-primary" (click)="next()">Next random question</button>
-
   <h2>{{question?.trainingContent}}</h2>
 
   <div class="mt-4 row">
@@ -52,4 +50,5 @@
     </form>
   </div>
 </div>
+
 <button class="btn btn-warning" (click)="back()">Back</button>
